added data types to environment variable settings
