+++ src/config/index.ts
@@ -1,23 +1,14 @@
+import { IConfig } from "../types/config";
+
 const Config: IConfig = {
   DATABASE: {
     URI: process.env.MONGODB_URI || "mongodb://localhost/photo-gallery",
-    USER: process.env.USER,
-    PASSWORD: process.env.PASSWORD,
+    USER: process.env.USER || "fernando",
+    PASSWORD: process.env.PASSWORD || "fernando",
   },
   PORT: process.env.PORT || 4000,
   NODE_ENV: process.env.NODE_ENV || "development",
+  JWT_KEY: process.env.JWT_KEY || "b8afrl7hkQeylazu",
 };
-
-interface IConfig {
-  DATABASE: Idbconfig;
-  PORT: string | number;
-  NODE_ENV: string;
-}
-
-interface Idbconfig {
-  URI: string;
-  USER: string | undefined;
-  PASSWORD: string | undefined;
-}
 
 export default Config;
